Fix remove stale hashes
1 - check if the path is for saving files/hashes was wrong
2 - the purge found erased labels - and didnt failed in assert.
if not exist in labels just remove.

+++ cpp/Osmosis/ObjectStore/Purge.cpp
@@ -21,8 +21,10 @@ void Purge::purge( boost::filesystem::path & dirToPurge )
 	size_t after = _staleHashes.size();
 	TRACE_INFO( "Purge found " << after << " objects to purge (" <<
 			( before - after ) << " remain)" );
-	for ( auto & hash : _staleHashes )
-		boost::filesystem::remove( _store.filenameForExisting( hash ) );
+	for ( auto & hash : _staleHashes ){
+		if ( _store.exists( hash ) )
+			boost::filesystem::remove( _store.filenameForExisting( hash ) );
+	}
 	BACKTRACE_END
 }
